Strip out the bcast part as a seperate executable.
It is actually more convenient this way. bcast can also be used to deploy
the main executable and relative packages.

I made this because importing from NERSC project on 4096 cores can still be
a pain. bcast allows me to package and deliver the program to /dev/shm.

Actually python-mpi-bcast can be replaced with bcast and some shell scripts.
I need to think about this more carefully.