Protect against giant mirror directories.
This was a problem invoked by having large catalogue files
in imaginglss script directory.

tar-dir will issue an error and quite if there is more than 50 MB of
code to be mirrored. This should be a reasonable upper limit of size for
on an agile, developing code base that doesn't work via pip.